Create util file for tests, run the tests in menuWorkExperience in parallel
